// src/app/core/services/model/eventTypes.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type EventTypes = 'ALARM' | 'EVENT' | 'DISRUPTION' | 'MAINTENANCE' | 'SYSTEM_EVENT' | 'SYSTEM_FAULT' | 'FUNCTION' | 'FATAL_PERSISTANT_SYSTEM_FAILURE';

export const EventTypes = {
    ALARM: 'ALARM' as EventTypes,
    EVENT: 'EVENT' as EventTypes,
    DISRUPTION: 'DISRUPTION' as EventTypes,
    MAINTENANCE: 'MAINTENANCE' as EventTypes,
    SYSTEMEVENT: 'SYSTEM_EVENT' as EventTypes,
    SYSTEMFAULT: 'SYSTEM_FAULT' as EventTypes,
    FUNCTION: 'FUNCTION' as EventTypes,
    FATALPERSISTANTSYSTEMFAILURE: 'FATAL_PERSISTANT_SYSTEM_FAILURE' as EventTypes
};