// src/app/core/services/model/reportDataQueryParameters.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OutputDataType } from './outputDataType';

/**
 * Parameters for report options query  The data query uses uuid instead of DevicePointName and channelChannelNo instead of ChannelName.  These two are keep to allow a quick change (data query) if needed.
 */
export interface ReportDataQueryParameters { 
    /**
     * uuid of the device point which created this data point
     */
    devicePointName?: string;
    /**
     * uuid of the device point which created this data point
     */
    devicePointUuid?: string;
    /**
     * name of the device, e.g. \"Wall-E 2\"
     */
    deviceUuid?: string;
    /**
     * type of the device, e.g. \"Wall-E\"
     */
    deviceTypeName?: string;
    /**
     * number of the channel, e.g. 6 (6 = Temperature_Get, see enum ChannelValueCodes or emico c api)
     */
    dataPointId?: number;
    /**
     * channel description, e.g. Get Temperature
     */
    dataPointDescription?: string;
    /**
     * channel unit, e.g. DegC
     */
    dataPointUnit?: string;
    dateUtcMin?: string;
    /**
     * Consider only datalog elements with date less or equal DateUTCMax
     */
    dateUtcMax?: string;
    dateUtcOffsetBase?: string;
    /**
     * Alternative option to set DateUTCMin and DateUTCMax  Only in combination with DateUTC
     */
    dateOffsetSec?: number;
    /**
     * Value should be greater equal ValueMin  if we want to allow negatives values, we have to change the generator constaint \"if (valueMin lower 0.0) then\"
     */
    valueMin?: number;
    /**
     * Value should be lower equal ValueMax
     */
    valueMax?: number;
    dataType?: OutputDataType;
    /**
     * should include virtual initial value  if no DataPoint matches filter criteria the last DataPoint befor selected timeframe is given
     */
    virtualInitValue?: boolean;
    /**
     * Limits the number of returned data log entries (obtained by the query)
     */
    limit: number;
    /**
     * Offset for the query result (useful for pagination)
     */
    offset?: number;
    /**
     * Predicate for the seek pagination (row id)
     */
    skipToken?: number;
    /**
     * Order the returned data log entries (by now there is only the \"Date\" column supported)
     */
    orderBy?: string;
    /**
     * Order of the returned data log entries (dependent to OrderBy)  \"asc\" or \"desc\" allowed, default OrderBy column is \"date\"
     */
    order?: string;
}