// src/app/core/services/model/sftpOptionsInput.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * sftp-server options
 */
export interface SftpOptionsInput { 
    /**
     * hostname of the sftp-server, e.g. 127.0.0.1
     */
    host: string;
    /**
     * port of the sftp-server, e.g. 22
     */
    port: number;
    /**
     * user name for the sftp-server login
     */
    username: string;
    /**
     * password for the sftp-server login
     */
    password: string;
    /**
     * path on the sftp-server (store new files on this location)
     */
    destinationPath: string;
    /**
     * fingerprint of sftp-server
     */
    fingerprint?: string;
    /**
     * path of public key
     */
    privateKeyPath?: string;
    /**
     * passphrase of key
     */
    passphrase?: string;
}